Sort by sensor name when doing KATCP -> TANGO translation
If you add sensors to a tango device in a random order, they also show up in the
GUI tools in that order, making it somewhat hard to find the sensor you are
looking for.

Of course this would get messed up again if the KATCP interface changes and new
sensors are added, but better for now.