Use EnsureOmniThread in TangoDevice2KatcpProxy
If we use more than one of the TangoDevice2KatcpProxy translators
in the same process, it doesn't work.  We see this in katproxy's
SkaMpiDshProxy if we add multiple TangoDeviceHandlers.  The
Tango event subsystem gets blocked.  This is a known issue when
using events from standard Python threads, and can be avoided
by using PyTango's EnsureOmniThread context handler.  That
feature was only added in PyTango v9.3.2.

JIRA: MT-1391