Use CHANGE_EVENT for inspecting client sampling (+hack)
Getting attribute samples via pushed events from the Tango device
server is much more efficient than polling. The server-side polling
can also degrade performance on the server (e.g., for the SPFC on
the MPI Dish prototype). Thus the default behaviour is now to leave
polling unmodified and just attempt to subscribe to the
`CHANGE_EVENT`.

Other types of events (archive, user, data ready) are no longer
considered either, as they

In order to detect successful subscriptions we have to use the
`stateless=False` option.  In testing, restarting the Tango
device after the proxy is up, most of the attribute readings
did not recover.  Not sure if that is related to this flag.

As a fallback option, server-side polling can be enabled.  We do
this as a hack for the MPI Dish master.

Tests have not been updated yet.

JIRA: MT-1391
(cherry picked from commit d7c9015402bcbd3cceea0428923a79404ea2addc)