Handle requests for multiple KATCP servers
The `_request_handlers` dict is created by the katcp `DeviceMetaclass`.
The result is a single dict for all instances of the server class.
If we run multiple servers in the same process, we need unique dicts
per instance.  So copy initial values into instance variable.
Same for the other dicts created by the metaclass.

In testing, some of the commands executed by the device proxies
are not returning `Future` objects even when green mode is set to
`Futures`.  If two servers are running, one of them returns futures
and the other does not.  Not sure why.

Tests have not been updated yet.

JIRA: MT-1391
(cherry picked from commit 3742bf1d75305c9e1964f4578b9846553ac285ee)