Encode latin-1 chars as utf-8 for KATCP
KATCP expects utf-8 encoding, so the description and units string
from Tango are converted from latin-1.

Arbitrary bytes in KATCP descriptions or units are not serialisable
via json, which katstore uses.  As an example, the degree symbol
is `'\xb0'` in latin-1 encoding causes problems with the json
library.

Tests have not been updated yet.

JIRA: MT-1391